requirements.txt
mujoco==3.8.0
yapf
pre-commit
websockets
pyyaml
Pillow
